refactor(preset): create config helpers
