+++ .gitignore
@@ -5,6 +5,7 @@ packages/*/node_modules
 
 packages/*/build
 packages/*/.tsup
+packages/*/styled-system
 
 tsconfig.tsbuildinfo
 */tsconfig.tsbuildinfo
