+++ docs/package.json
@@ -31,7 +31,6 @@
     "@mdx-js/react": "^3.1.0",
     "@next/mdx": "^15.2.0",
     "@pandacss/dev": "^0.53.0",
-    "@pandacss/preset-panda": "^0.53.0",
     "@pandacss/types": "^0.53.0",
     "@shikijs/rehype": "^3.1.0",
     "@types/node": "^22.13.8",
