+++ packages/panda-preset/package.json
@@ -22,6 +22,7 @@
     "@ark-ui/anatomy": "^3.5.0",
     "@cerberus-design/configs": "workspace:*",
     "@pandacss/dev": "^0.53.0",
+    "@pandacss/preset-panda": "^0.53.0",
     "@pandacss/types": "^0.53.0"
   },
   "publishConfig": {
