+++ packages/panda-preset/src/index.ts
@@ -1,56 +1,9 @@
-import {
-  defineConfig,
-  definePreset,
-  type Config,
-  type Preset,
-} from '@pandacss/dev'
-import { globalCss } from './globalCss'
-import { conditions } from './conditions'
-import { utilities } from './utilities'
-import { patterns } from './patterns'
-import { baseTheme, type RawThemes } from './theme'
-import { acheronTheme } from './themes/acheron'
-
 /**
  * This module contains the Cerberus preset and configuration options.
  * @module
  **/
 
-export const supportedThemes: RawThemes[] = ['cerberus', 'acheron']
-
-export const cerberusPreset: Preset = definePreset({
-  name: 'cerberus',
-
-  globalCss,
-  conditions,
-  utilities,
-  patterns,
-
-  // default theme: cerberus
-  theme: baseTheme,
-
-  // optional themes
-  themes: {
-    cerberus: baseTheme,
-    acheron: acheronTheme,
-  },
-
-  // opt-into additional theme variants
-  staticCss: {
-    themes: ['cerberus', 'acheron'],
-  },
-})
-
-export const cerberusConfig: Config = defineConfig({
-  preflight: true,
-  prefix: 'cerberus',
-
-  jsxFramework: 'react',
-  jsxFactory: 'cerberus',
-
-  outdir: 'styled-system',
-  importMap: '@cerberus/styled-system',
-})
+export * from './config'
 
 export * from './conditions'
 export * from './patterns'
