+++ packages/react/src/aria-helpers/nav-menu.aria.ts
@@ -8,6 +8,9 @@ interface NavTriggerAriaReturn {
   ['aria-expanded']: boolean
 }
 
+/**
+ * @deprecated will be removed in a future release
+ */
 export function createNavTriggerProps(
   values: NavTriggerAriaValues,
 ): NavTriggerAriaReturn {
