+++ packages/react/tsup.config.ts
@@ -7,10 +7,6 @@ const external = [
   ...Object.keys(peerDependencies),
   'react',
   'react-dom',
-  '@cerberus/styled-system/css',
-  '@cerberus/styled-system/patterns',
-  '@cerberus/styled-system/recipes',
-  '@cerberus/styled-system/jsx',
 ]
 
 export default defineConfig([
